refactor(test): move PublicClient
